modify styles & fix input component
